Use migration generator
